refactor: 구매할 수 있는 로또 티켓 개수를 구하는 기능 LotteryMoney로 위임 - Money -> LotteryMoney 클래스명 변경
