+++ src/main/java/lotterymachine/vo/LotteryMoney.java
@@ -2,23 +2,23 @@ package lotterymachine.vo;
 
 import java.util.Objects;
 
-public class Money {
+public class LotteryMoney {
     private static final int TICKET_PRICE = 1000;
     private static final String NOT_PURCHASABLE_EXCEPTION = "1000보다 큰 금액을 입력해야합니다.";
 
     private final int amount;
 
-    private Money(int amount) {
+    private LotteryMoney(int amount) {
         this.amount = amount;
     }
 
-    public static Money from(int amount) {
-        return new Money(amount);
+    public static LotteryMoney from(int amount) {
+        return new LotteryMoney(amount);
     }
 
-    public static Money fromInputAmount(int amount) {
+    public static LotteryMoney fromInputAmount(int amount) {
         validate(amount);
-        return new Money(amount);
+        return new LotteryMoney(amount);
     }
 
     private static void validate(int amount) {
@@ -27,12 +27,12 @@ public class Money {
         }
     }
 
-    public double divide(Money money) {
+    public double divide(LotteryMoney money) {
         return (double) this.amount / money.amount;
     }
 
-    public int getAmount() {
-        return amount;
+    public Count divideByLotteryPrice() {
+        return Count.from(this.amount / TICKET_PRICE);
     }
 
     @Override
@@ -43,7 +43,7 @@ public class Money {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        Money money = (Money) o;
+        LotteryMoney money = (LotteryMoney) o;
         return amount == money.amount;
     }
